+++ packages/ui/src/atoms/RadioGroup.stories.tsx
@@ -19,24 +19,36 @@ const render: Story["render"] = (args) => {
   return (
     <RadioGroup className="flex items-end gap-4 p-4" defaultValue="2">
       <div
-        className={cn("flex flex-col items-center gap-4", args.label ? "w-32" : "w-16")}
+        className={cn(
+          "flex flex-col items-center gap-4",
+          args.label ? "w-32" : "w-16",
+        )}
       >
         <RadioGroupItem {...args} value="1" />
       </div>
       <div
-        className={cn("flex flex-col items-center gap-4", args.label ? "w-32" : "w-16")}
+        className={cn(
+          "flex flex-col items-center gap-4",
+          args.label ? "w-32" : "w-16",
+        )}
       >
         <span className="text-grey-300 text-xs">:checked</span>
         <RadioGroupItem {...args} value="2" />
       </div>
       <div
-        className={cn("flex flex-col items-center gap-4", args.label ? "w-32" : "w-16")}
+        className={cn(
+          "flex flex-col items-center gap-4",
+          args.label ? "w-32" : "w-16",
+        )}
       >
         <span className="text-grey-300 text-xs">:disabled</span>
         <RadioGroupItem {...args} disabled={true} value="3" />
       </div>
       <div
-        className={cn("flex flex-col items-center gap-4", args.label ? "w-32" : "w-16")}
+        className={cn(
+          "flex flex-col items-center gap-4",
+          args.label ? "w-32" : "w-16",
+        )}
       >
         <span className="text-grey-300 text-xs">:focus</span>
         <RadioGroupItem {...args} autoFocus={true} value="4" />