+++ packages/ui/src/molecules/index.ts
@@ -2,3 +2,4 @@ export * from "./Carousel";
 export * from "./Form";
 export * from "./Pagination";
 export * from "./Tabs";
+export * from "./ScrollArea";